refactor(components): extract `RepositoryContent` into own component
to use it both in `ProjectCard` and `FeaturedProjectCard`
